% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\docType{data}
\name{comments}
\alias{comments}
\title{comments}
\format{
A data frame with 6 rows and 4 variables:
\describe{
\item{\code{topic}}{character feedback section}
\item{\code{choiceNames}}{character short display name for feedback}
\item{\code{choiceValues}}{character text to add to feedback}
\item{\code{selected}}{logical should checkbox be selected initially}
}
}
\usage{
comments
}
\description{
Sample comments
}
\details{
DETAILS
}
\keyword{datasets}
