% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedback.R
\name{feedback}
\alias{feedback}
\title{Feedback}
\usage{
feedback(com, ...)
}
\arguments{
\item{com}{Path to comment file or data.frame of comments}

\item{...}{Arguments to \code{readr::read_delim()} if \code{com} if the path to a file.}
}
\description{
\code{feedback()} opens a shiny gadget which complies selected feedback from a
file or data.frame, with the opportunity to write bespoke text.
When the feedback is complete, clip the copy button to copy the text,
and then the reset button.
}
\details{
the file (or data.frame) \code{com} needs to have four columns.
\itemize{
\item topic: collective name for group of comments. Need to be in the expected order.
\item choiceNames: short name displayed with the checkbox
\item choiceValues: text that is added to the feedback
\item selected: boolean specifying whether comment should be included by default.
}
}
\examples{
# feedback(comments)

# path to example file with feedback
fs::path_package("writeFeedback", "extdata/comments.csv")

}
